backend/requirements_all.txt
# Core FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
# psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI/ML Libraries
transformers==4.35.0
torch>=2.0.0
sentence-transformers==2.2.2
chromadb==0.4.22
langchain==0.1.0
langchain-community==0.0.13

# CNN and Computer Vision
tensorflow>=2.13.0
keras>=2.13.0
opencv-python==4.8.1.78
numpy>=1.24.0

# Notifications
twilio==8.10.0
sendgrid==6.10.0

# PDF Generation
reportlab==4.0.7
weasyprint==61.0

# Utilities
requests==2.31.0
pillow==10.1.0
python-dateutil==2.8.2

# Testing
pytest==7.4.3
httpx==0.25.1
pytest-asyncio==0.21.1
